+++ scripts/publish_crates.sh
@@ -37,7 +37,7 @@ publish_one() {
     return
   fi
   echo "Publishing $crate v$version"
-  cargo publish -p "$crate" --token "$CARGO_REGISTRY_TOKEN"
+  cargo publish -p "$crate"
   # allow index to update before dependents publish
   sleep 20
 }